OpeningCu Cut Scene and Main Menu
I added the opening cutscene Jade made and the buttons to the main menu.
I also did some of the opening cut scene script and also Jolie did some
of the scripting too we fixed some of the bugs but there is still one
but Jolie and me are paire programming to get it fixed. Also added main
menu sound Anthony made